Refactor LiDAR source by extracting UniformRaySource
