+++ sensorsim/lidars/__init__.py
@@ -1,0 +1,1 @@
+from .lidar import LiDAR, OusterLiDAR
