+++ sensorsim/light/__init__.py
@@ -1,0 +1,1 @@
+from .uniformRaySource import UniformRaySource
